requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
httpx
streamlit
pandas